adding controlFeedGen.ts to controll feed generators
