+++ .env.example
@@ -23,3 +23,22 @@ FEEDGEN_PUBLISHER_DID="did:plc:abcde...."
 
 # Delay between reconnect attempts to the firehose subscription endpoint (in milliseconds)
 FEEDGEN_SUBSCRIPTION_RECONNECT_DELAY=3000
+
+# Your handle name in bluesky
+YOUR_HANDLE=
+
+# A password for your handle
+YOUR_PASSWORD=
+
+# A short name for the record that will show in urls
+# Lowercase with no spaces. Ex: whats-hot
+FEED_RECORD_NAME=
+
+# A name displayed as feed name on bluesky clients
+FEED_DISPLAY_NAME=
+
+# A description about your feed
+FEED_DESCRIPTION=
+
+# A path to image file to be used as feed's avatar
+FEED_AVATAR_PATH=
